refactor: extract message_builder and serialization from agent.py
Sample PR demonstrating Option 1 (module functions + thin delegators) for
breaking down the 12,500-line agent.py into smaller, focused modules.

Extracted modules:
- message_builder.py (~1,850 lines): system/user message construction,
  knowledge retrieval, message formatting
- serialization.py (~840 lines): to_dict, from_dict, save, load, delete,
  deep_copy

Pattern: methods become standalone functions taking `agent: Agent` as first
param. Agent class keeps one-line delegators preserving the public API.

Result: agent.py goes from 12,524 to 10,021 lines (~20% reduction from
just these two modules). Full extraction of all groups would bring it to
~5,000 lines.
